cmd/insta/resources/data/argilla/requirements.txt
argilla
httpx
transformers[torch]
Pillow 